Add tests for channel upgrade completion
