+++ tools/integration-test/src/tests/channel_upgrade/mod.rs
@@ -1,3 +1,3 @@
 pub mod ics29;
-pub mod manual_channel_upgrade;
 pub mod upgrade_handshake;
+pub mod upgrade_handshake_steps;
